docker_images/build_model/requirements.txt
requests
numpy
scikit-surprise